cli/pyproject.toml
[project]
name = "finops-cli"
version = "0.1.0"
description = "FinOps Orchestrator CLI"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
]

[project.scripts]
finops-analyze = "main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "models", "api_client", "formatter"]

